some documentation added

+++ protected/controllers/AjaxController.php
@@ -1,12 +1,13 @@
 <?php
-
+/**
+ * Ajax requests controller
+ * Responsible for vhost removing
+ */
 class AjaxController extends Controller
 {
-    protected $_vhost;
-
-        /**
-         * @return array action filters
-         */
+    /**
+     * @return array action filters
+     */
     public function filters()
     {
         return array(
@@ -14,42 +15,60 @@ class AjaxController extends Controller
         );
     }
 
-       /**
-        * Specifies the access control rules.
-        * This method is used by the 'accessControl' filter.
-        * @return array access control rules
-        */
+    /**
+     * Specifies the access control rules.
+     * This method is used by the 'accessControl' filter.
+     * @return array access control rules
+     */
     public function accessRules()
     {
-            return array(
-                       array('allow',  // allow all users to access 'error' action.
-                               'actions'=>array('index'),
-                               'users'=>array('*'),
-                       ),
-                       array('allow', // allow authenticated users to access all actions
-                               'users'=>array('@'),
-                       ),
-                       array('deny',  // deny all users
-                               'users'=>array('*'),
-                       ),
+        return array(
+                array('allow',  // allow all users to access 'error' action.
+                    'actions'=>array('index'),
+                    'users'=>array('*'),
+                ),
+                array('allow', // allow authenticated users to access all actions
+                    'users'=>array('@'),
+                ),
+                array('deny',  // deny all users
+                    'users'=>array('*'),
+                ),
             );
     }
 
+    /**
+     * Default ajax page
+     * returns 404, as we don't need default ajax actions
+     * @throws CHttpException
+     */
     public function actionIndex()
    	{
         throw new CHttpException(404,'The requested page does not exist.');
    	}
 
     /**
-   	 * Remove domain
+   	 * Remove vhost record
    	 */
     public function actionRemoveVhost()
     {
-        $this->_loadVhost()->delete();
+        $vhost = null;
+        if(isset($_GET['vhostId'])) {
+            if(Yii::app()->user->isGuest) {
+                throw new CHttpException(404,'The requested page does not exist.');
+            }
+            $vhost = Vhost::model()->findByPk($_GET['vhostId'], 'owner_id=' . Yii::app()->user->getId());
+        }
+        // Return 404 if the vhost was not found
+        if($vhost === null) throw new CHttpException(404,'The requested page does not exist.');
+
+        $vhost->delete();
         echo json_encode(array('status'=>'OK','vhostId'=>$_GET['vhostId']));
         Yii::app()->end();
     }
 
+    /**
+   	 * Remove  a list of a vhost records
+   	 */
     public function actionRemVhostList()
     {
         if (isset($_POST['vhostToRemove'])) {
@@ -60,30 +79,13 @@ class AjaxController extends Controller
                     $removedVhosts[] = $vhostToRemove->id;
                     $vhostToRemove->delete();
                 }
-                $status = count($removedVhosts) == count($_POST['vhostToRemove']) ? 'OK' : 'PARTLY';
+                $status = (count($removedVhosts) == count($_POST['vhostToRemove'])) ? 'OK' : 'PARTLY';
                 echo json_encode(array('status'=>$status,'vhostIds'=>$removedVhosts));
                 Yii::app()->end();
             }
         }
-
+        // Return 404 by default
         throw new CHttpException(404,'The requested page does not exist.');
         Yii::app()->end();
-    }
-
-    protected function _loadVhost()
-    {
-        if($this->_vhost===null)
-      	{
-            if(isset($_GET['vhostId'])) {
-                if(Yii::app()->user->isGuest) {
-                    throw new CHttpException(404,'The requested page does not exist.');
-                }
-                $this->_vhost=Vhost::model()->findByPk($_GET['vhostId'], 'owner_id=' . Yii::app()->user->getId());
-            }
-
-            if($this->_vhost===null)
-                throw new CHttpException(404,'The requested page does not exist.');
-        }
-        return $this->_vhost;
     }
 }
